requirements.txt
livekit-agents>=0.8.0
livekit-plugins-cartesia>=0.6.0
livekit-plugins-deepgram>=0.6.0
livekit-plugins-openai>=0.9.0
python-dotenv>=1.0.1
